refactor: reorganize dialog package structure
