+++ app/src/androidTest/java/com/owncloud/android/ui/dialog/DialogFragmentIT.kt
@@ -38,9 +38,9 @@ import com.nextcloud.client.account.RegisteredUser
 import com.nextcloud.client.account.Server
 import com.nextcloud.client.device.DeviceInfo
 import com.nextcloud.client.documentscan.AppScanOptionalFeature
-import com.nextcloud.ui.ChooseAccountDialogFragment
-import com.nextcloud.ui.ChooseAccountDialogFragment.Companion.newInstance
-import com.nextcloud.ui.SetOnlineStatusBottomSheet
+import com.nextcloud.ui.dialog.account.ChooseAccountDialogFragment
+import com.nextcloud.ui.dialog.account.ChooseAccountDialogFragment.Companion.newInstance
+import com.nextcloud.ui.dialog.account.onlineStatus.SetOnlineStatusBottomSheet
 import com.nextcloud.ui.fileactions.FileActionsBottomSheet.Companion.newInstance
 import com.nextcloud.utils.EditorUtils
 import com.owncloud.android.AbstractIT
