+++ app/src/main/java/com/nextcloud/client/di/ComponentsModule.java
@@ -27,11 +27,11 @@ import com.nextcloud.client.widget.DashboardWidgetConfigurationActivity;
 import com.nextcloud.client.widget.DashboardWidgetProvider;
 import com.nextcloud.client.widget.DashboardWidgetService;
 import com.nextcloud.receiver.NetworkChangeReceiver;
-import com.nextcloud.ui.ChooseAccountDialogFragment;
-import com.nextcloud.ui.ChooseStorageLocationDialogFragment;
+import com.nextcloud.ui.dialog.account.ChooseAccountDialogFragment;
+import com.nextcloud.ui.dialog.chooseStorage.ChooseStorageLocationDialogFragment;
 import com.nextcloud.ui.ImageDetailFragment;
-import com.nextcloud.ui.SetOnlineStatusBottomSheet;
-import com.nextcloud.ui.SetStatusMessageBottomSheet;
+import com.nextcloud.ui.dialog.account.onlineStatus.SetOnlineStatusBottomSheet;
+import com.nextcloud.ui.dialog.account.statusMessage.SetStatusMessageBottomSheet;
 import com.nextcloud.ui.composeActivity.ComposeActivity;
 import com.nextcloud.ui.fileactions.FileActionsBottomSheet;
 import com.nextcloud.ui.trashbinFileActions.TrashbinFileActionsBottomSheet;
